bdm/processing/reddit/requirements.txt
pyspark==3.5.6
jsonschema==4.24.0
pandas==2.3.0
pytest>=8.4.0
python-magic==0.4.27
boto3==1.34.59
